Forced padded enc. symbols and fixed bugs:
Now encoding symbols always have T bytes, even the last source symbol
which includes padding in the buffer. This avoids complex code to check
different lengths when receiving source symbols.

Fixed a very well hidden bug (at least from me). Turns out that when you
wrap an array in a ByteBuffer, you have to be careful if you want the
buffer position to be always 0. You have to slice the array after
wrapping it. Took a while to figure it out.