Updated README (read rest of message):
Updated README file in master branch, but did not generate new release/tag. This should be avoided for other files. Typically, when code is sufficiently improved in the development branch, a merge is done on top of the master branch and a new release is created with incremented version number.