Added test for benchmark testing of buffer based methods:
The results indicate that using buffers is an order of magnitude slower
than using arrays of bytes.

The idea of reducing the number of XORs by performing word-by-word XORs
is not feasible because using buffers (even direct ones) creates a
non-negligible overhead.

Damn it Java. Y U no have pointers